予約内容変更メソッド作成（仮）
