+++ components/detailHotel/detailPlans.vue
@@ -82,7 +82,15 @@ export default {
     },
     preReserve() {
       const plan = this.plans[0];
+      // 予約情報IDの生成
+      let reserveId = "";
+      for (let i = 0; i < 7; i++) {
+        let num = Math.floor(Math.random() * 10) + 11;
+        let str = String(num);
+        reserveId += str;
+      }
       this.$store.commit("setPreReserveData", {
+        reserveId: reserveId,
         hotelName: this.basicInfo.hotelName,
         checkInDate: this.checkInDate,
         staySpan: this.staySpan,
@@ -95,7 +103,7 @@ export default {
         adultNum: this.adultNum,
       });
       console.log(this.$store.state.preReserveData);
-      this.$router.push("/reserveForm");
+      this.$router.push(`/reserveForm/${reserveId}`);
     },
   },
 };
