+++ components/reservetionContents.vue
@@ -50,7 +50,6 @@
 
 <script>
 import { format } from "date-fns";
-import { de } from "date-fns/locale";
 
 export default {
   name: "reservetionContents",
@@ -76,6 +75,8 @@ export default {
   },
   data() {
     return {
+      // 予約者ID
+      reserveId: 0,
       // ホテル名
       hotelName: "",
       // チェックイン日
@@ -125,7 +126,7 @@ export default {
   mounted() {
     let reserveDetail = {};
     reserveDetail = this.$store.getters.getPreReserveData;
-    console.log(reserveDetail);
+    this.reserveId = reserveDetail.reserveId;
     this.hotelName = reserveDetail.hotelName;
     this.date = reserveDetail.checkInDate;
     this.formatDate = format(new Date(this.date), "yyyy年MM月dd日");
@@ -279,17 +280,9 @@ export default {
       if (array.length > 0) {
         return;
       }
-      // 予約情報IDの生成
-      let reserveId = "";
-      for (let i = 0; i < 7; i++) {
-        let num = Math.floor(Math.random() * 10) + 11;
-        let str = String(num);
-        reserveId += str;
-      }
 
       // storeに送るためのオブジェクト生成(予約者情報)
       let object = {
-        reserveId: "",
         fullName1: "",
         fullName2: "",
         zipcode: "",
@@ -310,6 +303,7 @@ export default {
       };
       // storeに送るためのオブジェクト生成(プラン詳細)
       let detailObject = {
+        reserveId: 0,
         hotelName: "",
         formatDate: new Date(),
         staySpan: 0,
@@ -324,7 +318,6 @@ export default {
       };
 
       // 作ったオブジェクトに情報を代入する
-      object.reserveId = reserveId;
       object.fullName1 = this.fullName1;
       object.fullName2 = this.fullName2;
       object.zipcode = this.zipcode;
@@ -343,6 +336,7 @@ export default {
       object.card_name = this.card_name;
       object.other = this.other;
 
+      detailObject.reserveId = this.reserveId;
       detailObject.hotelName = this.hotelName;
       detailObject.formatDate = this.formatDate;
       detailObject.staySpan = this.staySpan;
