+++ pages/reserveForm/_id.vue
@@ -253,6 +253,8 @@ export default {
   },
   data() {
     return {
+      // 予約者ID
+      reserveId: 0,
       // フルネーム（氏名）のエラー
       fullName1Error: "",
       // フルネーム（かな）のエラー
@@ -338,6 +340,8 @@ export default {
    * 非同期処理.
    */
   mounted() {
+    // this.reserveId = Number(this.$route.params.id);
+    // console.log(resereveId);
     let reserveDetail = {};
     reserveDetail = this.$store.getters.getPreReserveData;
     this.adult = reserveDetail.adultNum;
